requirements.txt


flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.0.3
werkzeug==3.0.1
python-dateutil==2.8.2